- integrated snake into main menu
